refactoring notebooks and multimodal_rag module renamed
